+++ Frameworks/Adium Framework/Source/AIListGroupBubbleCell.m
@@ -29,7 +29,7 @@
 	if ((self = [super init]))
 	{
 		outlineBubble = NO;
-		outlineBubbleLineWidth = 1.0;
+		outlineBubbleLineWidth = 1.0f;
 		drawBubble = YES;
 	}
 	
@@ -37,10 +37,10 @@
 }
 
 //Give ourselves extra padding to compensate for the rounded bubble
-- (int)leftPadding{
+- (CGFloat)leftPadding{
 	return [super leftPadding] + EDGE_INDENT;
 }
-- (int)rightPadding{
+- (CGFloat)rightPadding{
 	return [super rightPadding] + EDGE_INDENT;
 }
 
@@ -75,9 +75,9 @@
 	if ([self cellIsSelected]) {
 		NSColor *highlightColor = [controlView highlightColor];
 		NSGradient 	*gradient = (highlightColor ?
-								 [[[NSGradient alloc] initWithStartingColor:highlightColor endingColor:[highlightColor darkenAndAdjustSaturationBy:0.4]] autorelease] :
+								 [[[NSGradient alloc] initWithStartingColor:highlightColor endingColor:[highlightColor darkenAndAdjustSaturationBy:0.4f]] autorelease] :
 								 [NSGradient selectedControlGradient]);
-		[gradient drawInBezierPath:[NSBezierPath bezierPathWithRoundedRect:[self bubbleRectForFrame:cellFrame]] angle:270.0];
+		[gradient drawInBezierPath:[NSBezierPath bezierPathWithRoundedRect:[self bubbleRectForFrame:cellFrame]] angle:270.0f];
 	}
 }
 
@@ -88,7 +88,7 @@
 		NSBezierPath	*bezierPath;
 		
 		bezierPath = [NSBezierPath bezierPathWithRoundedRect:[self bubbleRectForFrame:inRect]];
-		[[self backgroundGradient] drawInBezierPath:bezierPath angle:90.0];
+		[[self backgroundGradient] drawInBezierPath:bezierPath angle:90.0f];
 		
 		if (outlineBubble) {
 			[bezierPath setLineWidth:outlineBubbleLineWidth];
@@ -113,18 +113,18 @@
 	
 	//Margin for the drop highlight
 	rect.size.width -= DROP_HIGHLIGHT_WIDTH_MARGIN;
-	rect.origin.x += DROP_HIGHLIGHT_WIDTH_MARGIN / 2.0;
+	rect.origin.x += DROP_HIGHLIGHT_WIDTH_MARGIN / 2.0f;
 	
 	rect.size.height -= DROP_HIGHLIGHT_HEIGHT_MARGIN;
-	rect.origin.y += DROP_HIGHLIGHT_HEIGHT_MARGIN / 2.0;	
+	rect.origin.y += DROP_HIGHLIGHT_HEIGHT_MARGIN / 2.0f;	
 	
 	NSBezierPath *path = [NSBezierPath bezierPathWithRoundedRect:[self bubbleRectForFrame:rect]];
 
-	[[[NSColor blueColor] colorWithAlphaComponent:0.2] set];
+	[[[NSColor blueColor] colorWithAlphaComponent:0.2f] set];
 	[path fill];
 	
-	[[[NSColor blueColor] colorWithAlphaComponent:0.8] set];
-	[path setLineWidth:2.0];
+	[[[NSColor blueColor] colorWithAlphaComponent:0.8f] set];
+	[path setLineWidth:2.0f];
 	[path stroke];
 	
 	[NSGraphicsContext restoreGraphicsState];
